Update resources-available.markdown
Removed the hyperthreading stuff, as we never communicate this to users. It is only confusing, because they do not need to do or can do anything with that information.

Also, we have 4 GPU devices available in the Grid K2's and Ceph storage has a maximum of 675, due to the fact that 25% of the technical capacity cannot be used in order to maintain system stability.