New note @ 'Fetching the contextualization files'
A user reported that pointing the "Path in OpenNebula server" to the contextualization file was not successful (https://helpdesk.surfsara.nl/ticket/41216). 

Check & confirmed ... SysAdmins also do not know the reason.
To mitigate the issue, adding an extra note to the documentation. 
A better approach will be to review the section and re-write it ...